// src/types/vacancy.ts
export type IVacancy = {
    id: string;
    username: string;
    status: string;
    hh_id: string;
    name: string;
    area: string;
    salary_from: string;
    salary_to: string;
    address: string;
    url: string;
    company: string;
    requirements: string;
    responsobility: string;
    scedule: string;
    role: string;
    experience: string;
}